maintainability: decompose gitea/tracker_client.go by responsibility (#830)
tracker_client.go was 741 lines, ~59 from the 800-line hard cap
(scripts/file_size_budget_test.go, empty baseline) with 21 commits in 4
weeks — split it pre-emptively before a feature PR is forced into a rushed
mid-stream split (#830, the 2-of-2 follow-up to #821).

Move by responsibility seam (no renames, no behavior change), keeping the
client core (struct, constructors, wire type Issue, accessors, issue-number
cache) in tracker_client.go:
- tracker_client_listing.go — ListIssuesByStates + label/state paging + scoping + cap probe.
- tracker_client_refresh.go — FetchIssueStatesBy{IDs,Refs} narrow state refresh + ref/number helpers + the single-issue fetch primitive (shared with blocker resolution).
- tracker_client_blocker.go — Depends-on parsing + per-tick blocker cache + blocker-ref building.
- tracker_client_mapping.go — time/id/label/state-query mapping helpers + diagnostics + state set.

Verified lossless: the top-level declaration set is byte-identical
before/after (46 decls) and the non-import body line multiset is identical
(673 lines); each new file's import block is the goimports-trimmed subset
its group uses. tracker_client.go 741 -> 110; every file is now well under
the 800 cap (largest sibling 268). No public API added for the budget;
package gitea is unchanged externally.

Second of #821's two files (its own PR per the acceptance criteria); the
first (tracker/github.go) landed in #831.

Closes #830
